tools/law_library/requirements.txt
chromadb>=0.5.20
anthropic>=0.40.0
openai>=1.54.0
python-dotenv>=1.0.0
pypdf>=5.1.0
python-docx>=1.1.2
sentence-transformers>=5.1.0
regex>=2024.0.0